+++ lib/requestcore/requestcore.class.php
@@ -161,6 +161,11 @@ class RequestCore
 	 */
 	public $allow_set_time_limit = true;
 
+	/**
+	 * Whether or not to use gzip encoding via CURLOPT_ENCODING
+	 */
+	public $use_gzip_enconding = true;
+
 
 	/*%******************************************************************************************%*/
 	// CONSTANTS
@@ -686,7 +691,7 @@ class RequestCore
 		}
 
 		// Handle the encoding if we can.
-		if (extension_loaded('zlib'))
+		if ($this->use_gzip_enconding && extension_loaded('zlib'))
 		{
 			curl_setopt($curl_handle, CURLOPT_ENCODING, 'gzip, deflate');
 		}