+++ utilities/request.class.php
@@ -65,6 +65,11 @@ class CFRequest extends RequestCore
 		$this->credentials = $credentials;
 		$this->cacert_location = ($this->credentials['certificate_authority'] ? $this->credentials['certificate_authority'] : false);
 
+		if (strpos(parse_url($url, PHP_URL_HOST), 'dynamodb') === 0)
+			{
+				$this->use_gzip_enconding = false;
+			}
+
 		return $this;
 	}
 }